New Year
Decided to make this, will come back to clean it up. Right now it's just a cute little ugly program that makes a search on google for "Bender Quotes", looks at the contents of the first 10 results, processes them with Word2Vec, then looks at the distance between "Shiny" and "Metal". Turns out they're pretty close, looks like a success so far. Next up: Clean the code, make it read a sentence and jumble it up with words relevant to the search terms. How? Oh there's so many ways with Word2Vec, thinking maybe pick a pool of 20 potential word replacements, whichever is closest to the one before will be the next word in the sentence to jumble, then so on until there aren't any words, first word maybe stay the same?